requirements.txt
telethon
openai
python-dotenv
langdetect
tqdm
loguru
